+++ project/src/main/java/com/tqs/project/service/BusinessService.java
@@ -1,10 +1,10 @@
-package com.tqs.project.Service;
+package com.tqs.project.service;
 
 import java.util.List;
 import java.util.Optional;
 
-import com.tqs.project.Model.Business;
-import com.tqs.project.Repository.BusinessRepository;
+import com.tqs.project.model.Business;
+import com.tqs.project.repository.BusinessRepository;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
