+++ project/src/test/java/com/tqs/project/service/DeliveryServiceTest.java
@@ -13,10 +13,9 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
 
-import com.tqs.project.Model.Courier;
-import com.tqs.project.Model.Delivery;
-import com.tqs.project.Repository.DeliveryRepository;
-import com.tqs.project.Service.DeliveryService;
+import com.tqs.project.model.Courier;
+import com.tqs.project.model.Delivery;
+import com.tqs.project.repository.DeliveryRepository;
 
 import static org.assertj.core.api.Assertions.assertThat;
