+++ project/src/test/java/com/tqs/project/service/ShopServiceTest.java
@@ -10,18 +10,14 @@ import org.mockito.internal.verification.VerificationModeFactory;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.util.Arrays;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Optional;
 
-import com.tqs.project.Model.Address;
-import com.tqs.project.Model.Business;
-import com.tqs.project.Model.Delivery;
-import com.tqs.project.Model.Shop;
+import com.tqs.project.model.Address;
+import com.tqs.project.model.Business;
+import com.tqs.project.model.Shop;
 
-import com.tqs.project.Repository.ShopRepository;
-
-import com.tqs.project.Service.ShopService;
+import com.tqs.project.repository.ShopRepository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -38,10 +34,10 @@ import static org.assertj.core.api.Assertions.assertThat;
     @BeforeEach
     public void setUp() {
 
-        Shop coviran = new Shop("Coviran", new Address(), new Business(), new HashSet<Delivery>());
+        Shop coviran = new Shop("Coviran", new Address(), new Business());
         coviran.setId(111L);
 
-        Shop pingodoce = new Shop("Pingo Doce", new Address(), new Business(), new HashSet<Delivery>());
+        Shop pingodoce = new Shop("Pingo Doce", new Address(), new Business());
 
         List<Shop> allShops = Arrays.asList(coviran, pingodoce);
 
@@ -78,8 +74,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 
     @Test
      void whengetAll_thenReturn2Records() {
-        Shop coviran = new Shop("Coviran", new Address(), new Business(), new HashSet<Delivery>());
-        Shop pingodoce = new Shop("Pingo Doce", new Address(), new Business(), new HashSet<Delivery>());
+        Shop coviran = new Shop("Coviran", new Address(), new Business());
+        Shop pingodoce = new Shop("Pingo Doce", new Address(), new Business());
 
         List<Shop> allShops = service.getAllShops();
