+++ project/src/test/java/com/tqs/project/service/UserServiceTest.java
@@ -13,9 +13,8 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
 
-import com.tqs.project.Model.User;
-import com.tqs.project.Repository.UserRepository;
-import com.tqs.project.Service.UserService;
+import com.tqs.project.model.User;
+import com.tqs.project.repository.UserRepository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -83,6 +82,6 @@ import static org.assertj.core.api.Assertions.assertThat;
         // verify if FindAllUsers is called once
         Mockito.verify(rep, VerificationModeFactory.times(1)).findAll();
 
-        assertThat(allUsers).hasSize(4).extracting(User::getUsername).contains(alex.getUsername(), john.getUsername(), peter.getUsername(), gon.getUsername());
+        assertThat(allUsers).hasSize(4).extracting(User::getEmail).contains(alex.getEmail(), john.getEmail(), peter.getEmail(), gon.getEmail());
     }
  }
